Merged the two existing projects into one
Merged the xApi and LrsClient projects into one project to simplify the usage and creation
of the NuGet package.

It doesn't make a lot of sense to separate the two projects since it would mean creating two different
NugGet packages and I don't think people would want to use one without the other anyway.

It can always be separated again in the future if need be.