Removed the HttpClient as Singleton in LrsClient
Decided to not leave the HttpClient instance as a Singleton in the LrsClient class. This made the class awkward to use. I'll document that the best use of the LrsClient is as a Singleton for users of the class, but this leave the users the possibility to manage the lifecycle of the LrsClient more easily.